requirements.txt
python-binance
pyTelegramBotAPI
requests